+++ src/Contracts/MessageProducer.php
@@ -50,6 +50,10 @@ interface MessageProducer extends InteractsWithConfigCallbacks
     /** Set Sasl configuration. */
     public function withSasl(string $username, string $password, string $mechanisms, string $securityProtocol = 'SASL_PLAINTEXT'): self;
 
+    public function withFlushRetries(int $retries): self;
+    
+    public function withFlushTimeout(int $timeoutInMs): self;
+
     /** Specifies which serializer should be used. */
     public function usingSerializer(MessageSerializer $serializer): self;