+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All relevant changes to `mateusjunges/laravel-kafka` will be documented here.
 
+##[2024-11-22 v2.4.1](https://github.com/mateusjunges/laravel-kafka/compare/v2.4.0...v2.4.1)
+* Allow to customize flush settings (retries + timeout) by @sash in [#317](https://github.com/mateusjunges/laravel-kafka/pull/317)
+
 ##[2024-11-04 v2.4.0](https://github.com/mateusjunges/laravel-kafka/compare/v2.3.2...v2.4.0)
 * Allow to use other types of Keys other than strings by @mateusjunges in [#322](https://github.com/mateusjunges/laravel-kafka/pull/322)